feat: Se agregaron las pruebas unitarias del router del modulo de jugadores
Pruebas unitarias de los modulos del repository
Se realizo un cambio minimo en models para que permita paremotros de los IDs
Cambio en repository para normalizar el item que se busca por ser "case-sensitive matching"

+++ backend/src/modules/players/models.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, func, ForeignKey, Index, text, JSON
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, func, ForeignKey, Index, text, JSON
 from config.database import Base
 
 class Player(Base):
     __tablename__ = "players"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     name = Column(String(128), nullable=False)
     position = Column(String(64), nullable=False)
     image_url = Column(String(512), nullable=True)
@@ -12,7 +12,7 @@ class Player(Base):
     is_active = Column(Boolean, nullable=False, default=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     created_by = Column(Integer, ForeignKey("users.id", ondelete="RESTRICT"), nullable=False)
-    team_id = Column(BigInteger, ForeignKey("teams.id", ondelete="CASCADE"), nullable=False)
+    team_id = Column(Integer, ForeignKey("teams.id", ondelete="CASCADE"), nullable=False)
 
 # Unique within a team (case-insensitive name)
 Index(
@@ -28,8 +28,8 @@ Index("ix_players_team_id", "team_id")
 class PlayerNews(Base):
     __tablename__ = "player_news"
 
-    id = Column(BigInteger, primary_key=True, index=True)
-    player_id = Column(BigInteger, ForeignKey("players.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, index=True)
+    player_id = Column(Integer, ForeignKey("players.id", ondelete="CASCADE"), nullable=False)
     author_id = Column(Integer, ForeignKey("users.id", ondelete="RESTRICT"), nullable=False)
     summary = Column(String(30), nullable=False)
     text = Column(String(300), nullable=False)
